// server/web/frontend/src/generated/proto/policy_ui_types.ts
// AUTO-GENERATED by protoc-gen-bor-ui. DO NOT EDIT.

export type PolicyFieldType =
  | "boolean"
  | "string"
  | "integer"
  | "string-enum"
  | "integer-enum"
  | "list"
  | "json"
  | "object";

export interface PolicySubFieldDef {
  key: string;
  label: string;
  description: string;
  type: Exclude<PolicyFieldType, "object" | "integer-enum" | "json">;
  stringOptions?: string[];
}

export interface PolicyFieldDef {
  key: string;
  label: string;
  group: string;
  description: string;
  type: PolicyFieldType;
  stringOptions?: string[];
  intOptions?: { value: number; label: string }[];
  subFields?: PolicySubFieldDef[];
  chromeOnly?: boolean;
}